Allow extra fields in scoring requests. (#71) (#72)
Cherry-pick of PR: #71 (1540ec4) into `support-1.6` LTS branch.

This has been requested in: https://github.com/h2oai/h2oai/issues/9395.
I don't think that we want to allow subset of fields in the requests as it may silently ignore typos in the input data (we already saw an example of that with the local scorer). So keeping that side of the check intact.

Also duplicated here: #70.