Restructure Gradle files (#106)
This is a follow up on @orendain comments in: https://github.com/h2oai/dai-deployment-templates/pull/105.

Turns out some of the comments cannot be implemented without having a lot of code copy&pasted around. E.g. `plugins` cannot be used anywhere except as the first part of a (sub)project gradle file. I want to prevent copying the shared configuration around as it is then hard to keep in sync. The result is that I still use the old way of enabling plugins where needed.

However the comments lead me to try a different way of structuring the gradle files. Please have a look and comment heavily. Maybe you won't find it better, in which case I am happy to drop the PR.

Things I personally like now are:
- it is possible to look at the subproject gradle file and see where stuff is defined (as the `apply from:` now work as sort of includes).
- separate stuff is in separate files, making the root gradle file much simpler.
- no need for central KDB related exclusions as this is now achieved by the kdb subprojects not using the style mixins (explicitely via `aplyt from: ... _no_style.gradle`).